fix: settings view feat: add separate views for settings
