+++ app/Livewire/SettingsBackup.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Livewire\Settings;
+namespace App\Livewire;
 
 use App\Jobs\DatabaseBackupJob;
 use App\Models\InstanceSettings;
@@ -10,7 +10,7 @@ use App\Models\Server;
 use App\Models\StandalonePostgresql;
 use Livewire\Component;
 
-class Backup extends Component
+class SettingsBackup extends Component
 {
     public InstanceSettings $settings;
 
@@ -41,8 +41,24 @@ class Backup extends Component
 
     public function mount()
     {
-        $this->backup = $this->database?->scheduledBackups->first() ?? null;
-        $this->executions = $this->backup?->executions ?? [];
+        if (isInstanceAdmin()) {
+            $settings = InstanceSettings::get();
+            $database = StandalonePostgresql::whereName('coolify-db')->first();
+            $s3s = S3Storage::whereTeamId(0)->get() ?? [];
+            if ($database) {
+                if ($database->status !== 'running') {
+                    $database->status = 'running';
+                    $database->save();
+                }
+                $this->database = $database;
+            }
+            $this->settings = $settings;
+            $this->s3s = $s3s;
+            $this->backup = $this->database?->scheduledBackups?->first() ?? null;
+            $this->executions = $this->backup?->executions ?? [];
+        } else {
+            return redirect()->route('dashboard');
+        }
     }
 
     public function add_coolify_database()
